+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][unreleased]
+## Added
+- Added test to confirm `EventDispatcher` is exported by the package main.
+- Added test to confirm that `Event` is attached to `EventDispatcher`.
+- Added `priority` property to `EventDispatcher` with seven presets.
 
 ## [0.1.1][2015-07-09]
 ### Fixed